+++ src/Pregunta.Me/Pregunta.Me.Services/Administration/Experts/IExpertRegistrationRequest.cs
@@ -13,7 +13,7 @@ namespace Pregunta.Me.Services.Administration
         string LastName { get; set; }
         string Email { get; set; }
         string Currency { get; set; }
-        decimal BillingRate { get; set; }
+        double BillingRate { get; set; }
         string Language { get; set; }
         string Country { get; set; }
     }
